pyproject.toml
[project]
name = "2-weather-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.3.0",
    "openai-agents>=0.0.7",
    "python-dotenv>=1.1.0",
]
